+++ android/App/src/main/kotlin/com/darakeon/dfm/welcome/WelcomeActivity.kt
@@ -27,6 +27,8 @@ class WelcomeActivity : BaseActivity() {
 			action_close.applyGlyphicon()
 
 			MovesService.start(this)
+
+			populateCache()
 		} else {
 			redirect<LoginActivity>()
 		}