+++ android/Lib/src/main/kotlin/com/darakeon/dfm/lib/auth/SP.kt
@@ -1,6 +1,7 @@
 package com.darakeon.dfm.lib.auth
 
 import android.content.Context
+import com.google.gson.Gson
 
 private const val sharedPreferencesDFM = "DfM"
 
@@ -14,6 +15,12 @@ fun Context.setValue(key: String, value: String) {
 		apply()
 	}
 }
+
+inline fun <reified T> Context.getValueTyped(key: String): T =
+	Gson().fromJson(this.getValue(key), T::class.java)
+
+inline fun <reified T> Context.setValueTyped(key: String, value: T) =
+	this.setValue(key, Gson().toJson(value))
 
 private val Context.sp
 	get() = getSharedPreferences(