+++ docs/RELEASES.md
@@ -15,7 +15,7 @@ Legend:
 - :ant:: the developer is improving stuff or fixing some sh\*t it did (increases last number)
 
 ## <a name="6.2.0.0"></a>6.2.0.0 :whale: <sup>`1`</sup>
-- [ ] Cache account/category at login + refresh button at app
+- [ ] `210716>......` Cache account/category at login + refresh button at app
 
 ## <a name="6.1.0.0"></a>6.1.0.0 :dragon: <sup>`3`</sup>
 - [x] `210714>210715` Add commit linter